# sers-project/lib/api-zod/src/generated/types/emotionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERS - System for Emotional Response and Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface EmotionRecord {
  id: number;
  studentId: number;
  sessionId: number;
  emotion: string;
  /** green | yellow | red */
  alertLevel: string;
  /** @nullable */
  confidence?: number | null;
  recordedAt: string;
}