# sers-project/lib/api-zod/src/generated/types/emotionRecordInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERS - System for Emotional Response and Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface EmotionRecordInput {
  studentId: number;
  sessionId: number;
  emotion: string;
  alertLevel: string;
  confidence?: number;
}