# sers-project/lib/api-zod/src/generated/types/parentNotification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERS - System for Emotional Response and Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface ParentNotification {
  id: number;
  studentId: number;
  message: string;
  emotion: string;
  emailSent: boolean;
  /** @nullable */
  emailAddress?: string | null;
  sentAt: string;
}