# sers-project/lib/api-zod/src/generated/types/parentNotificationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERS - System for Emotional Response and Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface ParentNotificationInput {
  studentId: number;
  /** @minLength 1 */
  message: string;
  emotion: string;
}