# sers-project/lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERS - System for Emotional Response and Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  classroomId: number;
  name: string;
  /** @nullable */
  studentCode?: string | null;
  /** @nullable */
  parentEmail?: string | null;
  /** @nullable */
  parentToken?: string | null;
  createdAt: string;
}