# sers-project/lib/api-zod/src/generated/types/studentState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERS - System for Emotional Response and Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from "./student";

export interface StudentState {
  student: Student;
  /** @nullable */
  latestEmotion: string | null;
  /** @nullable */
  alertLevel: string | null;
  /** @nullable */
  lastUpdated?: string | null;
}