# sers-project/lib/api-zod/src/generated/types/studentTrends.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERS - System for Emotional Response and Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { EmotionCount } from "./emotionCount";
import type { EmotionRecord } from "./emotionRecord";

export interface StudentTrends {
  studentId: number;
  emotionCounts: EmotionCount[];
  recentRecords: EmotionRecord[];
  /** @nullable */
  dominantEmotion?: string | null;
  /** @nullable */
  alertLevel?: string | null;
}